lorentzian_strategy/requirements.txt
# Core Data Science and ML Libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
numba>=0.57.0

# Financial Data Analysis
vectorbt>=0.25.0
ta-lib>=0.4.25
yfinance>=0.2.0
quantlib>=1.31

# Performance and Optimization
cython>=0.29.0
bottleneck>=1.3.0
numexpr>=2.8.0

# Configuration and Data Management
pydantic>=2.0.0
pyyaml>=6.0
toml>=0.10.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
psutil>=5.9.0
memory-profiler>=0.61.0

# Optional GPU Support
# cupy-cuda11x>=12.0.0  # Uncomment if using GPU
# cudf>=23.08.0         # Uncomment for GPU DataFrames